WEB-3582: move server load positions to own file to fix build error
