+++ app/(candidate)/dashboard/content/page.js
@@ -5,8 +5,7 @@ import candidateAccess from '../shared/candidateAccess';
 import ContentPage from './components/ContentPage';
 import { fetchUserCampaign } from 'app/(candidate)/onboarding/shared/getCampaign';
 import { getServerUser } from 'helpers/userServerHelper';
-import { serverLoadCandidatePosition } from 'app/(candidate)/dashboard/campaign-details/components/issues/issuesUtils';
-
+import { serverLoadCandidatePosition } from 'app/(candidate)/dashboard/campaign-details/components/issues/serverLoadCandidatePosition';
 const meta = pageMetaData({
   title: 'Campaign Content | GoodParty.org',
   description: 'Campaign Content',
