+++ app/(candidate)/dashboard/questions/page.js
@@ -2,11 +2,8 @@ import { fetchUserCampaign } from 'app/(candidate)/onboarding/shared/getCampaign
 import pageMetaData from 'helpers/metadataHelper';
 import candidateAccess from '../shared/candidateAccess';
 import QuestionsPage from './components/QuestionsPage';
-import {
-  fetchIssues,
-  serverLoadCandidatePosition,
-} from 'app/(candidate)/dashboard/campaign-details/components/issues/issuesUtils';
-
+import { fetchIssues } from 'app/(candidate)/dashboard/campaign-details/components/issues/issuesUtils';
+import { serverLoadCandidatePosition } from 'app/(candidate)/dashboard/campaign-details/components/issues/serverLoadCandidatePosition';
 const meta = pageMetaData({
   title: 'Additional Questions | GoodParty.org',
   description: 'Additional Questions',
